+++ node/events/updateGlovoOrderStatus.ts
@@ -53,6 +53,7 @@ export async function updateGlovoOrderStatus(ctx: StatusChangeContext) {
       message: `Glovo order ${glovoPayload.glovoOrderId} status update failed`,
       status: 500,
       payload: error,
+      error,
     })
   }
 }