+++ node/utils/orders.ts
@@ -9,8 +9,10 @@ export const createVtexOrderData = (
   clientProfileData: ClientProfileData,
   marketplace: boolean
 ): CreateOrderPayload => {
-  const { order_id, estimated_total_price } = glovoOrder
-  const { items, pickupPoints, postalCode, logisticsInfo } = orderSimulation
+  const { order_id } = glovoOrder
+  const { items, pickupPoints, postalCode, logisticsInfo, totals } =
+    orderSimulation
+
   const {
     email,
     firstName,
@@ -64,10 +66,15 @@ export const createVtexOrderData = (
     []
   )
 
+  const totalValue = totals.reduce(
+    (total: number, item: SimulationTotalsItem) => (total += item.value),
+    0
+  )
+
   const vtexOrderData: CreateOrderPayload = {
     marketplaceOrderId: order_id,
     marketplaceServicesEndpoint: 'https://api.glovoapp.com/',
-    marketplacePaymentValue: estimated_total_price,
+    marketplacePaymentValue: totalValue,
     marketplaceOrderGroup: order_id,
     isCreatedAsync: true,
     items: updatedItems,
@@ -112,8 +119,8 @@ export const createVtexOrderData = (
         {
           installments: 1,
           paymentSystem: glovoOrder.payment_method,
-          referenceValue: glovoOrder.estimated_total_price,
-          value: glovoOrder.estimated_total_price,
+          referenceValue: totalValue,
+          value: totalValue,
         },
       ],
     }